requirements.txt
mtcnn~=0.1.1
numpy~=1.22.3
opencv-python~=4.5.5.64
tensorflow
scipy~=1.8.0
Pillow~=9.1.0
tqdm~=4.64.0
torch~=1.11.0
keras~=2.8.0